% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{linreg}
\usage{
linreg(y, x, method = "qr", intercept = T, CI_level = 0.95)
}
\arguments{
\item{y}{a column vector, treated as outcome.}

\item{x}{a numerical matrix, treated as predictor.}

\item{method}{Character value, "qr" (by default), "svd" or "inverse"}

\item{intercept}{logical, "TRUE" by default.
If "FALSE", the model will not include the intercept.}

\item{CI_level}{Numeric number, 0.95 (by default).
Should be in (0,1) and directly change alpha used in confidence interval.}
}
\value{
A list contains regression results.
}
\description{
fit simple regression models and multiple regression models.
}
\examples{
 x <- matrix(c(-1, -2, -3, -4, -5, 1, 3, 5, 7, 9), nrow = 5)
 y <- c(2, 5, 8, 11, 19)
 # fit the linear model without an intercept
 linreg(y, x, intercept = FALSE)

}
