% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pexp.R
\name{my_pexp}
\alias{my_pexp}
\title{Exponential CDF}
\usage{
my_pexp(x, rate = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{a numeric quantile}

\item{rate}{a numeric value of rate parameter (default 1)}

\item{lower.tail}{a logical value indicating whether to compute
P[X<=x] if TRUE or P[X>x] if FALSE (default TRUE)}

\item{log.p}{a logical value indicating whether to return
the log probability (default FALSE)}
}
\value{
cumulative distribution function evaluated at x
}
\description{
Calculates the cumulative probability of the exponential distribution function

my_pexp() functions identically to the pexp() function in the 'stats' package of R
for scalar value of x (non-vectorized)
}
