% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{square}
\alias{square}
\title{Hello, world!}
\usage{
square(s)
}
\arguments{
\item{s}{input value}
}
\value{
the square of x
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!'.
And I added a square function in it.
}
\examples{
square(3)

}
